+++ Package.swift
@@ -21,6 +21,7 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/danielsaidi/BadgeIcon.git", .upToNextMajor(from: "2.1.4")),
         .package(url: "https://github.com/simibac/ConfettiSwiftUI.git", .upToNextMajor(from: "2.0.3")),
+        .package(url: "https://github.com/danielsaidi/ObservablePersistency.git", .upToNextMajor(from: "0.1.0")),
         .package(url: "https://github.com/danielsaidi/OnboardingKit.git", .upToNextMajor(from: "9.0.0")),
         .package(url: "https://github.com/danielsaidi/PresentationKit.git", .upToNextMajor(from: "0.5.0")),
         .package(url: "https://github.com/danielsaidi/ScanCodes.git", .upToNextMajor(from: "0.2.0")),
@@ -34,6 +35,7 @@ let package = Package(
             dependencies: [
                 "BadgeIcon",
                 "ConfettiSwiftUI",
+                "ObservablePersistency",
                 "OnboardingKit",
                 "PresentationKit",
                 "ScanCodes",